Adding Solution Code and removing Application code.
