--- Antheias_Garden/AntheiasKotlinSolution/.idea/.gitignore
@@ -1,3 +1,0 @@
-# Default ignored files
-/shelf/
-/workspace.xml
