--- Antheias_Garden/AntheiasKotlinSolution/app/src/main/java/com/eaglewarrior/antheiaskotlinsolution/MainActivity.kt
@@ -1,34 +1,0 @@
-package com.eaglewarrior.antheiaskotlinsolution
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import java.time.temporal.ValueRange
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
-
-fun main() {
-
-    val numbOfLily = 1.0
-    val daysUntilFullMoon = 29
-    var totalNumOfLily = numbOfLily
-
-    for (i in 1..daysUntilFullMoon) {
-        println("Day ${i}, There are ${totalNumOfLily} Lilies")
-        totalNumOfLily = totalNumOfLily * 2
-    }
-
-
-    var lilyNum = 1
-    var daysNum = 1
-
-    while (daysNum <= 29) {
-        println("On day ${daysNum}, there should be ${lilyNum} Lilies in the pond.")
-        lilyNum = lilyNum * 2
-        daysNum++
-    }
-}
